remove: Instance and Witness types for CCS & CCCS
This PR moves a tep forward to the NIMFS-centric API.
First, removes all the `_Witness` & `_Instance` structs for CCS & CCCS
as they added a bunch of overhead.

Sedonc, it also re-assembles internal fm APIs so that they ask for
slices instead of refs to `Vec`.

Finally, reassembles all the tests to indeed modify CCS & CCCS
interaction as described avobe.